% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cck94v2-package.R
\docType{package}
\name{cck94-package}
\alias{cck94}
\alias{cck94-package}
\title{cck94: Replication of 'Optimal Fiscal Policy in a Business Cycle Model' by Chari, Christiano, and Kehoe (1994)}
\description{
Replication of Chari et al. (1994) using projections (original method), time iteration, the Endogenous Grid Method (ECM) by Maliar and Maliar (2013), and separate validation by using Ramsey Optimal Taxation Using a Monte Carlo 1st-order Linear Quadratic (LQ) approximation and a 2nd-order perturbation (Benigno and Woodford, 2005).
}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
