% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cck94_params}
\alias{cck94_params}
\title{NSSS and LQ diagnostics}
\usage{
cck94_params(args)
}
\description{
Returns steady-state shares, BW utility coefficients, and
  LQ loss function coefficients for diagnostic purposes.
}
