% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cck94_solve_ti}
\alias{cck94_solve_ti}
\title{Solve using Time Iteration}
\usage{
cck94_solve_ti(args, n_k, n_z, n_g, max_iter, tol, T_total, T_burn, seed)
}
\arguments{
\item{args}{Model parameters (same as cck94_stats)}

\item{n_k}{Number of capital grid points (default 50)}

\item{n_z}{Number of technology shock states (default 5)}

\item{n_g}{Number of govt spending shock states (default 5)}

\item{max_iter}{Maximum iterations (default 500)}

\item{tol}{Convergence tolerance (default 1e-6)}

\item{T_total}{Simulation periods for statistics (default 500000)}

\item{T_burn}{Burn-in periods (default 60000)}

\item{seed}{Random seed (default 42)}
}
\description{
Solves the CCK94/BW Ramsey taxation model using time iteration
  (policy function iteration) on the Euler equation.
}
