% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp4r.R
\name{cck94_stats_2nd}
\alias{cck94_stats_2nd}
\title{BW Table 6: 2nd-order Monte Carlo statistics}
\usage{
cck94_stats_2nd(args, n_k, T_total, T_burn, seed)
}
\description{
Simulates second-order Ramsey policy rules using SGU (2004)
  perturbation and returns the same statistics as Table 6 of Benigno &
  Woodford (2005).
}
